requirements.txt
streamlit
pandas
PyMuPDF
pdfplumber
faiss-cpu
numpy
sentence-transformers
langchain
langchain-huggingface
huggingface_hub